requirements.txt
python-telegram-bot==13.7
pandas==1.3.5
matplotlib==3.4.3
numpy==1.21.2
